Feat : shared 페이지 구현
