+++ src/styles/card.css
@@ -1,10 +1,11 @@
 .card-button {
-    width: 340px;
-    height: 334px;
+    width: 100%;
+    height: 100%;
     box-shadow: 0px 5px 25px 0px rgba(0, 0, 0, 0.08);
     padding: 0;
     border: 0;
     background-color: transparent;
+    border-radius: 15px;
 }
 
 .card-button:hover {
@@ -22,15 +23,17 @@
 }
 
 .card-img-area {
+    width: 100%;
+    height: 200px;
+    flex-shrink: 0;
     overflow: hidden;
     border-radius: 15px 15px 0px 0px;
     background: #d9d9d9;
 }
 
 .card-img {
-    display: block;
-    border-radius: 15px 15px 0 0;
-    object-fit: cover;
+    width: 100%;
+    height: 100%;
 }
 
 .card-time {
