+++ src/styles/gnb.css
@@ -23,6 +23,14 @@ header {
 }
 
 @media screen and (max-width: 1200px) {
+    header {
+        display: flex;
+        padding: 33px 32.5px 32px 32.5px;
+        flex-direction: column;
+        align-items: center;
+        gap: 8px;
+        align-self: stretch;
+    }
     .header-container {
         width: 100%;
     }
