+++ src/styles/searchBar.css
@@ -1,0 +1,20 @@
+.searchbar-container {
+    display: flex;
+    width: 100%;
+    padding: 15px 16px;
+    align-items: flex-start;
+    gap: 10px;
+    border-radius: 10px;
+    background: #f5f5f5;
+}
+
+.searchbar-container img {
+    width: 16px;
+    height: 16px;
+}
+
+.searchbar-container input {
+    flex-grow: 1;
+    background-color: transparent;
+    border: none;
+}
